requirements.txt
autograd==1.8.0
autograd-gamma==0.5.0
build==1.2.2.post1
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cramjam==2.10.0
cycler==0.12.1
docutils==0.21.2
ecos==2.0.14
fastparquet==2024.11.0
fonttools==4.58.1
formulaic==1.1.1
fsspec==2025.5.1
gamma-pytools==2.1.3
id==1.5.0
idna==3.10
interface-meta==1.3.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
joblib==1.5.1
keyring==25.6.0
kiwisolver==1.4.8
lifelines==0.30.0
markdown-it-py==3.0.0
matplotlib==3.10.3
mdurl==0.1.2
mizani==0.13.5
more-itertools==10.7.0
mypy_extensions==1.1.0
narwhals==1.41.0
nh3==0.2.21
numexpr==2.10.2
numpy==1.26.4
osqp==0.6.7.post3
packaging==25.0
pandas==2.2.3
patsy==1.0.1
pillow==11.2.1
plotnine==0.14.5
Pygments==2.19.1
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
qdldl==0.1.7.post5
rdata==0.11.2
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
scikit-learn==1.6.1
scikit-lego==0.9.5
scikit-survival==0.24.1
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
sklearn-compat==0.1.3
statsmodels==0.14.4
threadpoolctl==3.6.0
twine==6.1.0
typing-inspect==0.9.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wrapt==1.17.2
xarray==2025.4.0
xlrd==2.0.1
